+++ R/as_terms.R
@@ -11,15 +11,40 @@
 #' data(markers)
 #' as_terms(markers)
 as_terms <- function(x, names = NULL, ...) {
+
     val <- validate_term_count(x)
     if (isTRUE(val)) {
+        term.vars <- attributes(x)[["term.vars"]]
         if (is.null(names)) names <- paste2(x[, attributes(x)[["group.vars"]], drop=FALSE], ...)
         x <- x[, attributes(x)[["term.vars"]], drop=FALSE]
     } else {
          if (is.null(names)) names <- rownames(x)
+         term.vars <- NULL
     }
     nms <- colnames(x)
 	lst <- apply(x, 1, function(y) rep(nms, y))
-    if(!is.list(lst)) lst <- lapply(1:ncol(lst), function(i) lst[, i])	
-    stats::setNames(lst, nm = names)
+    if(!is.list(lst)) lst <- lapply(1:ncol(lst), function(i) lst[, i])
+    out <- stats::setNames(lst, nm = names)
+
+    class(out) <- c('as_terms', class(out))
+
+    attributes(out)[["term.vars"]] <- term.vars
+    out
 }
+
+
+
+#' Prints an as_terms Object
+#'
+#' Prints an as_terms object.
+#'
+#' @param x The as_terms object.
+#' @param \ldots ignored
+#' @method print as_terms
+#' @export
+print.as_terms <- function(x, ...) {
+    x <- rm_class(x, "as_terms")
+    attributes(x)[["term.vars"]] <- NULL
+    print(x)
+}
+